Bugfixes, UI Desing&Testing, startter StateMachine implemented

+++ ABBModbusProj/inc/DecimalEdit.h
@@ -8,7 +8,7 @@
 
 class DecimalEdit : public PropertyEdit {
 public:
-	DecimalEdit(LiquidCrystal *lcd_, std::string editTitle, float _step, float _lowerLim, float _upperLim);
+	DecimalEdit(LiquidCrystal *lcd_, std::string editTitle, double _step, double _lowerLim, double _upperLim);
 	virtual ~DecimalEdit();
 	void increment();
 	void decrement();
@@ -17,21 +17,21 @@ public:
 	void setFocus(bool focus);
 	bool getFocus();
 	void display();
-	float getValue();
-	void setValue(float value);
-	void setStep(float _step);
+	double getValue();
+	void setValue(double value);
+	void setStep(double _step);
 protected:
 	bool save();
 	void displayEditValue();
 	LiquidCrystal *lcd;
 	std::string title;
-	float value;
-	float edit;
-	float step;
-	float lowerLim;
-	float upperLim;
+	double value;
+	double edit;
+	double step;
+	double lowerLim;
+	double upperLim;
 	bool focus;
-	static constexpr float tolerance = 0.001;
+	static constexpr double tolerance = 0.001;
 };
 
 #endif /* DECIMALEDIT_H_ */
